feat: add custom hex formatting
adds a new `hex_format` string option to the template frontmatter.

this string, if set, is used to render all hex-format colours in the
template with tera.

the following context variables are available:

- r: red
- g: green
- b: blue
- a: alpha
- z: alpha, or blank if alpha == 0xFF

these are all lowercase 2-digit hex values, for example '7e'.

all five of these have uppercase counterparts, for example `G`, which
are uppercase 2-digit hex values, for example '7E'.

the default format is `{{r}}{{g}}{{b}}{{z}}`.

Co-authored-by: Ruby <rubyowo284@pm.me>